[feat] API 추가

+++ src/main/java/com/mybury/waver/domain/Keyword.java
@@ -20,23 +20,23 @@ import lombok.NoArgsConstructor;
 @Builder
 public class Keyword {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Integer id;
 
-    @Builder.Default
-    @Enumerated(value = EnumType.STRING)
-    @Column(length = 10, nullable = false)
-    private KeywordType type = KeywordType.KEYWORD;
+  @Builder.Default
+  @Enumerated(value = EnumType.STRING)
+  @Column(length = 10, nullable = false)
+  private KeywordType type = KeywordType.KEYWORD;
 
-    private String name;
+  private String name;
 
-    @Builder.Default
-    @Column(nullable = false)
-    private Integer seq = 0;
+  @Builder.Default
+  @Column(nullable = false)
+  private Integer seq = 0;
 
-    @Builder.Default
-    @Enumerated(value = EnumType.STRING)
-    @Column(length = 1, nullable = false)
-    private YesNo recommendYn = YesNo.N;
+  @Builder.Default
+  @Enumerated(value = EnumType.STRING)
+  @Column(length = 1, nullable = false)
+  private YesNo recommendYn = YesNo.N;
 }
